Implement IconView to iOS
